Implement authorization based on user permissions.

+++ src/Application/Roles/Commands/DeleteRole/DeleteRole.cs
@@ -1,5 +1,6 @@
 ﻿using Forum.Application.Common.Interfaces;
 using Forum.Application.Common.Models;
+using Forum.Domain.Common;
 using MediatR;
 
 namespace Forum.Application.Roles.Commands.DeleteRole;
@@ -13,6 +14,9 @@ public class DeleteRoleCommandHandler(IRoleManager roleManager) : IRequestHandle
 {
     public async Task<CustomResponse> Handle(DeleteRoleCommand command, CancellationToken cancellationToken)
     {
+        if (command.RoleName == DefaultRoles.ADMIN)
+            return new CustomResponse() { Succeeded = false, Message = "Default admin role cannot be deleted" };
+
         return await roleManager.DeleteRoleAsync(command.RoleName, cancellationToken);
     }
 }
